+++ Estudo-Java/ProgramaCalculador/src/main/java/programacalculador/ProgramaCalculador.java
@@ -1,0 +1,18 @@
+package programacalculador;
+
+/**
+ * Programa que faz cálculos diversos para fins de estudo da tecnologia Java e 
+ * ferramenta JUnit 5.
+ */
+public class ProgramaCalculador {
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String[] args) {
+
+    MenuPrincipal menuPrincipal = new MenuPrincipal();
+    menuPrincipal.iniciar();
+    }
+    
+}
