--- Estudo-Java/ProgramaCalculador/src/programacalculador/ProgramaCalculador.java
@@ -1,17 +1,0 @@
-package programacalculador;
-
-/**
- * Programa que faz cálculos diversos para fins de estudo da tecnologia Java.
- */
-public class ProgramaCalculador {
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-
-    MenuPrincipal menuPrincipal = new MenuPrincipal();
-    menuPrincipal.iniciar();
-    }
-    
-}
